```text feat: Add authentication for applicants
This commit adds the `authApplicant` function to the `ApplicantStoreState` interface in `applicantStore.ts`. The `authApplicant` function is responsible for authenticating an applicant by making a POST request to the server with the applicant's number and date of birth. If the authentication is successful (status code 201), the applicant is logged in and the response data is stored in the `applicant` state. If the authentication fails (status code 200), an error message is displayed.

The purpose of this change is to provide authentication functionality for applicants in the application.